refactor(typechecker): extract ScopeStack into separate module
Moves ScopeStack implementation into its own file to improve separation of concerns and maintainability.
