+++ forst/pkg/typechecker/scope.go
@@ -61,61 +61,6 @@ func (s *Scope) LookupType(name ast.Identifier) (Symbol, bool) {
 	return Symbol{}, false
 }
 
-// Manages scopes during type checking, maintaining a tree of nested scopes
-type ScopeStack struct {
-	scopes  map[NodeHash]*Scope
-	current *Scope
-	Hasher  *StructuralHasher
-}
-
-func NewScopeStack(hasher *StructuralHasher) *ScopeStack {
-	globalScope := &Scope{
-		Symbols: make(map[ast.Identifier]Symbol),
-	}
-	return &ScopeStack{
-		scopes:  make(map[NodeHash]*Scope),
-		current: globalScope,
-		Hasher:  hasher,
-	}
-}
-
-func (ss *ScopeStack) PushScope(node ast.Node) {
-	hash := ss.Hasher.HashNode(node)
-	scope := &Scope{
-		Parent:   ss.current,
-		Node:     node,
-		Symbols:  make(map[ast.Identifier]Symbol),
-		Children: make([]*Scope, 0),
-	}
-	ss.current.Children = append(ss.current.Children, scope)
-	ss.current = scope
-	ss.scopes[hash] = scope
-}
-
-func (ss *ScopeStack) PopScope() {
-	if ss.current.Parent != nil {
-		ss.current = ss.current.Parent
-	}
-}
-
-func (ss *ScopeStack) CurrentScope() *Scope {
-	return ss.current
-}
-
-func (ss *ScopeStack) FindScope(node ast.Node) *Scope {
-	hash := ss.Hasher.HashNode(node)
-	return ss.scopes[hash]
-}
-
-// GlobalScope returns the root scope by traversing up the scope tree
-func (ss *ScopeStack) GlobalScope() *Scope {
-	scope := ss.current
-	for scope.Parent != nil {
-		scope = scope.Parent
-	}
-	return scope
-}
-
 type Symbol struct {
 	Identifier ast.Identifier
 	Types      []ast.TypeNode
